requirements.txt
gradio
google-generativeai
python-dotenv
youtube-transcript-api
langcodes